Implement interrupts for 12864 LCD encoder with RAMPS 1.4 configuration
A configuration item named USING_12864_WITH_ENCODER_INTERRUPTS
was added to the board_ramps.h include file to enable interrupt usage.

These changes require adding a jumper from D33 to D18, and one from
D31 to D19.  These can be done from the display connector to pins on
the RAMPS 1.4 board.  I cut a female jumper cable in half, soldered
the wire to the display adapter, and used the female end to attach to
the correct pin on the RAMPS 1.4 board.

D33 and D31 will still be set to inputs and receive the signal, but
they are not used.  Interrupts are enabled for D18 and D19, their
pull-ups are enabled, and the signal is read from the new pins.

Code was written to process the interrupts and handle menus.  Turning
the knob quicker will increment float and long values quicker.

The button press was not moved to an interrupt.  It still must be held
sufficiently long for the polling code to recognize the button press.